+++ example/src/examples/TakeSnapshot.tsx
@@ -9,9 +9,7 @@ import {
 } from 'react-native';
 
 import MapView, {Marker} from 'react-native-maps';
-// @ts-ignore
 import flagBlueImg from './assets/flag-blue.png';
-// @ts-ignore
 import flagPinkImg from './assets/flag-pink.png';
 
 const {width, height} = Dimensions.get('window');