+++ src/MapView.tsx
@@ -8,7 +8,6 @@ import {
   NativeSyntheticEvent,
   Platform,
   requireNativeComponent,
-  UIManager,
   ViewProps,
 } from 'react-native';
 import {
@@ -170,6 +169,12 @@ export type MapViewProps = ViewProps & {
    * @platform Android: Supported
    */
   liteMode?: boolean;
+
+  /**
+   * https://developers.google.com/maps/documentation/get-map-id
+   * google cloud mapId to enable cloud styling and more
+   */
+  googleMapId?: string;
 
   /**
    * Sets loading background color.
@@ -728,14 +733,7 @@ class MapView extends React.Component<MapViewProps, State> {
     this._onChange = this._onChange.bind(this);
   }
 
-  /**
-   * @deprecated Will be removed in v2.0.0, as setNativeProps is not a thing in fabric.
-   * See https://reactnative.dev/docs/new-architecture-library-intro#migrating-off-setnativeprops
-   */
   setNativeProps(props: Partial<NativeProps>) {
-    console.warn(
-      'setNativeProps is deprecated and will be removed in next major release',
-    );
     // @ts-ignore
     this.map.current?.setNativeProps(props);
   }
@@ -1046,6 +1044,8 @@ class MapView extends React.Component<MapViewProps, State> {
         initialRegion: null,
         onChange: this._onChange,
         onMapReady: this._onMapReady,
+        liteMode: this.props.liteMode,
+        googleMapId: this.props.googleMapId,
         ref: this.map,
         customMapStyleString: this.props.customMapStyle
           ? JSON.stringify(this.props.customMapStyle)
@@ -1067,6 +1067,8 @@ class MapView extends React.Component<MapViewProps, State> {
       props = {
         style: this.props.style,
         region: null,
+        liteMode: this.props.liteMode,
+        googleMapId: this.props.googleMapId,
         initialRegion: this.props.initialRegion || null,
         initialCamera: this.props.initialCamera,
         ref: this.map,
@@ -1077,14 +1079,6 @@ class MapView extends React.Component<MapViewProps, State> {
           ? JSON.stringify(this.props.customMapStyle)
           : undefined,
       };
-    }
-
-    if (Platform.OS === 'android' && this.props.liteMode) {
-      return (
-        <ProviderContext.Provider value={this.props.provider}>
-          <AIRMapLite {...props} />
-        </ProviderContext.Provider>
-      );
     }
 
     const AIRMap = getNativeMapComponent(this.props.provider);
@@ -1115,10 +1109,6 @@ if (Platform.OS === 'android') {
 }
 const getNativeMapComponent = (provider: Provider) =>
   airMaps[provider || 'default'];
-
-const AIRMapLite = UIManager.getViewManagerConfig('AIRMapLite')
-  ? requireNativeComponent<NativeProps>('AIRMapLite')
-  : () => null;
 
 export const AnimatedMapView = RNAnimated.createAnimatedComponent(MapView);